+++ src/soc/bus/range.rs
@@ -10,17 +10,17 @@ pub enum RangeKind {
 
 #[derive(Debug, Clone)]
 pub struct BusRange {
-    pub id: u64,
-    pub bus_start: u64,
-    pub bus_end: u64,
-    pub device_offset: u64,
+    pub id: usize,
+    pub bus_start: usize,
+    pub bus_end: usize,
+    pub device_offset: usize,
     pub device_id: usize,
     pub priority: u8,
     pub kind: RangeKind,
 }
 
 impl BusRange {
-    pub fn contains(&self, addr: u64) -> bool {
+    pub fn contains(&self, addr: usize) -> bool {
         self.bus_start <= addr && addr < self.bus_end
     }
 
@@ -28,7 +28,7 @@ impl BusRange {
         self.bus_start < other.bus_end && other.bus_start < self.bus_end
     }
 
-    pub fn len(&self) -> u64 {
+    pub fn len(&self) -> usize {
         self.bus_end - self.bus_start
     }
 
@@ -40,15 +40,15 @@ impl BusRange {
 #[derive(Clone)]
 pub struct ResolvedRange {
     pub device: Arc<dyn Device>,
-    pub bus_start: u64,
-    pub bus_end: u64,
-    pub device_offset: u64,
+    pub bus_start: usize,
+    pub bus_end: usize,
+    pub device_offset: usize,
     pub priority: u8,
     pub device_id: usize,
 }
 
 impl ResolvedRange {
-    pub fn len(&self) -> u64 {
+    pub fn len(&self) -> usize {
         self.bus_end - self.bus_start
     }
 
@@ -56,7 +56,7 @@ impl ResolvedRange {
         self.bus_start == self.bus_end
     }
 
-    pub fn contains(&self, addr: u64) -> bool {
+    pub fn contains(&self, addr: usize) -> bool {
         self.bus_start <= addr && addr < self.bus_end
     }
 }
